requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
fastapi==0.67.0
future==0.18.2
h11==0.12.0
idna==3.2
Jinja2==3.0.1
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
pydantic==1.8.2
pymongo==3.12.0
requests==2.26.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
